# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/vnd.cyclonedx+json; version=1.6"
    ],
    "swagger": "2.0",
    "info": {
        "description": "CBOMkit-theia analyzes cryptographic assets in a container image or directory.",
        "title": "CBOMkit-theia",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/image/get": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.cyclonedx+json; version=1.6"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Generate CBOM from existing image",
                "parameters": [
                    {
                        "description": "Request body containing the image identifier, list of activated plugins and BOM.",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.imageGetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cyclonedx.BOM"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cyclonedx.Advisory": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "cyclonedx.Annotation": {
            "type": "object",
            "properties": {
                "annotator": {
                    "$ref": "#/definitions/cyclonedx.Annotator"
                },
                "bom-ref": {
                    "type": "string"
                },
                "subjects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "text": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "cyclonedx.Annotator": {
            "type": "object",
            "properties": {
                "component": {
                    "$ref": "#/definitions/cyclonedx.Component"
                },
                "individual": {
                    "$ref": "#/definitions/cyclonedx.OrganizationalContact"
                },
                "organization": {
                    "$ref": "#/definitions/cyclonedx.OrganizationalEntity"
                },
                "service": {
                    "$ref": "#/definitions/cyclonedx.Service"
                }
            }
        },
...
        "cyclonedx.BOM": {
            "type": "object",
            "properties": {
                "$schema": {
                    "description": "JSON specific fields",
                    "type": "string"
                },
                "annotations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Annotation"
                    }
                },
                "bomFormat": {
                    "type": "string"
                },
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Component"
                    }
                },
                "compositions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Composition"
                    }
                },
                "declarations": {
                    "$ref": "#/definitions/cyclonedx.Declarations"
                },
                "definitions": {
                    "$ref": "#/definitions/cyclonedx.Definitions"
                },
                "dependencies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Dependency"
                    }
                },
                "externalReferences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.ExternalReference"
                    }
                },
                "formulation": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Formula"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/cyclonedx.Metadata"
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Property"
                    }
                },
                "serialNumber": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Service"
                    }
                },
                "specVersion": {
                    "$ref": "#/definitions/cyclonedx.SpecVersion"
                },
                "version": {
                    "type": "integer"
                },
                "vulnerabilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Vulnerability"
                    }
                }
            }
        },
...
        "cyclonedx.Component": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "Deprecated: Use authors or manufacturer instead.",
                    "type": "string"
                },
                "authors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.OrganizationalContact"
                    }
                },
                "bom-ref": {
                    "type": "string"
                },
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Component"
                    }
                },
                "copyright": {
                    "type": "string"
                },
                "cpe": {
                    "type": "string"
                },
                "cryptoProperties": {
                    "$ref": "#/definitions/cyclonedx.CryptoProperties"
                },
                "data": {
                    "$ref": "#/definitions/cyclonedx.ComponentData"
                },
                "description": {
                    "type": "string"
                },
                "evidence": {
                    "$ref": "#/definitions/cyclonedx.Evidence"
                },
                "externalReferences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.ExternalReference"
                    }
                },
                "group": {
                    "type": "string"
                },
                "hashes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Hash"
                    }
                },
                "licenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.LicenseChoice"
                    }
                },
                "manufacturer": {
                    "$ref": "#/definitions/cyclonedx.OrganizationalEntity"
                },
                "mime-type": {
                    "type": "string"
                },
                "modelCard": {
                    "$ref": "#/definitions/cyclonedx.MLModelCard"
                },
                "modified": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "omniborId": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pedigree": {
                    "$ref": "#/definitions/cyclonedx.Pedigree"
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Property"
                    }
                },
                "publisher": {
                    "type": "string"
                },
                "purl": {
                    "type": "string"
                },
                "releaseNotes": {
                    "$ref": "#/definitions/cyclonedx.ReleaseNotes"
                },
                "scope": {
                    "$ref": "#/definitions/cyclonedx.Scope"
                },
                "supplier": {
                    "$ref": "#/definitions/cyclonedx.OrganizationalEntity"
                },
                "swhid": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "swid": {
                    "$ref": "#/definitions/cyclonedx.SWID"
                },
                "type": {
                    "$ref": "#/definitions/cyclonedx.ComponentType"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "cyclonedx.ComponentData": {
            "type": "object",
            "properties": {
                "bom-ref": {
                    "type": "string"
                },
                "classification": {
                    "type": "string"
                },
                "contents": {
                    "$ref": "#/definitions/cyclonedx.ComponentDataContents"
                },
                "description": {
                    "type": "string"
                },
                "governance": {
                    "$ref": "#/definitions/cyclonedx.DataGovernance"
                },
                "graphics": {
                    "$ref": "#/definitions/cyclonedx.ComponentDataGraphics"
                },
                "name": {
                    "type": "string"
                },
                "sensitiveData": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "$ref": "#/definitions/cyclonedx.ComponentDataType"
                }
            }
        },
...
        "cyclonedx.Composition": {
            "type": "object",
            "properties": {
                "aggregate": {
                    "$ref": "#/definitions/cyclonedx.CompositionAggregate"
                },
                "assemblies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bom-ref": {
                    "type": "string"
                },
                "dependencies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "vulnerabilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "cyclonedx.CryptoProperties": {
            "type": "object",
            "properties": {
                "algorithmProperties": {
                    "$ref": "#/definitions/cyclonedx.CryptoAlgorithmProperties"
                },
                "assetType": {
                    "$ref": "#/definitions/cyclonedx.CryptoAssetType"
                },
                "certificateProperties": {
                    "$ref": "#/definitions/cyclonedx.CertificateProperties"
                },
                "oid": {
                    "type": "string"
                },
                "protocolProperties": {
                    "$ref": "#/definitions/cyclonedx.CryptoProtocolProperties"
                },
                "relatedCryptoMaterialProperties": {
                    "$ref": "#/definitions/cyclonedx.RelatedCryptoMaterialProperties"
                }
            }
        },
...
        "cyclonedx.Declarations": {
            "type": "object",
            "properties": {
                "affirmation": {
                    "$ref": "#/definitions/cyclonedx.Affirmation"
                },
                "assessors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Assessor"
                    }
                },
                "attestations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Attestation"
                    }
                },
                "claims": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Claim"
                    }
                },
                "evidence": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.DeclarationEvidence"
                    }
                },
                "signature": {
                    "$ref": "#/definitions/cyclonedx.JSFSignature"
                },
                "targets": {
                    "$ref": "#/definitions/cyclonedx.Targets"
                }
            }
        },
        "cyclonedx.Definitions": {
            "type": "object",
            "properties": {
                "standards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.StandardDefinition"
                    }
                }
            }
        },
        "cyclonedx.Dependency": {
            "type": "object",
            "properties": {
                "dependsOn": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ref": {
                    "type": "string"
                }
            }
        },
...
        "cyclonedx.Evidence": {
            "type": "object",
            "properties": {
                "callstack": {
                    "$ref": "#/definitions/cyclonedx.Callstack"
                },
                "copyright": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Copyright"
                    }
                },
                "identity": {
                    "$ref": "#/definitions/cyclonedx.EvidenceIdentity"
                },
                "licenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.LicenseChoice"
                    }
                },
                "occurrences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.EvidenceOccurrence"
                    }
                }
            }
        },
...
        "cyclonedx.ExternalReference": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "hashes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Hash"
                    }
                },
                "type": {
                    "$ref": "#/definitions/cyclonedx.ExternalReferenceType"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "cyclonedx.Formula": {
            "type": "object",
            "properties": {
                "bom-ref": {
                    "type": "string"
                },
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Component"
                    }
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Property"
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Service"
                    }
                },
                "workflows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Workflow"
                    }
                }
            }
        },
        "cyclonedx.Hash": {
            "type": "object",
            "properties": {
                "alg": {
                    "$ref": "#/definitions/cyclonedx.HashAlgorithm"
                },
                "content": {
                    "type": "string"
                }
            }
        },
...
        "cyclonedx.LicenseChoice": {
            "type": "object",
            "properties": {
                "expression": {
                    "type": "string"
                },
                "license": {
                    "$ref": "#/definitions/cyclonedx.License"
                }
            }
        },
...
        "cyclonedx.MLModelCard": {
            "type": "object",
            "properties": {
                "bom-ref": {
                    "type": "string"
                },
                "considerations": {
                    "$ref": "#/definitions/cyclonedx.MLModelCardConsiderations"
                },
                "modelParameters": {
                    "$ref": "#/definitions/cyclonedx.MLModelParameters"
                },
                "quantitativeAnalysis": {
                    "$ref": "#/definitions/cyclonedx.MLQuantitativeAnalysis"
                }
            }
        },
...
        "cyclonedx.Metadata": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.OrganizationalContact"
                    }
                },
                "component": {
                    "$ref": "#/definitions/cyclonedx.Component"
                },
                "licenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.LicenseChoice"
                    }
                },
                "lifecycles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Lifecycle"
                    }
                },
                "manufacture": {
                    "description": "Deprecated: Use Component Manufacturer instead.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cyclonedx.OrganizationalEntity"
                        }
                    ]
                },
                "manufacturer": {
                    "$ref": "#/definitions/cyclonedx.OrganizationalEntity"
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Property"
                    }
                },
                "supplier": {
                    "$ref": "#/definitions/cyclonedx.OrganizationalEntity"
                },
                "timestamp": {
                    "type": "string"
                },
                "tools": {
                    "$ref": "#/definitions/cyclonedx.ToolsChoice"
                }
            }
        },
...
        "cyclonedx.OrganizationalContact": {
            "type": "object",
            "properties": {
                "bom-ref": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "cyclonedx.OrganizationalEntity": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/cyclonedx.PostalAddress"
                },
                "bom-ref": {
                    "type": "string"
                },
                "contact": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.OrganizationalContact"
                    }
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "cyclonedx.Pedigree": {
            "type": "object",
            "properties": {
                "ancestors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Component"
                    }
                },
                "commits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Commit"
                    }
                },
                "descendants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Component"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "patches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Patch"
                    }
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Component"
                    }
                }
            }
        },
...
        "cyclonedx.Property": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "cyclonedx.SWID": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "patch": {
                    "type": "boolean"
                },
                "tagId": {
                    "type": "string"
                },
                "tagVersion": {
                    "type": "integer"
                },
                "text": {
                    "$ref": "#/definitions/cyclonedx.AttachedText"
                },
                "url": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "cyclonedx.Scope": {
            "type": "string",
            "enum": [
                "excluded",
                "optional",
                "required"
            ],
            "x-enum-varnames": [
                "ScopeExcluded",
                "ScopeOptional",
                "ScopeRequired"
            ]
        },
...
        "cyclonedx.Service": {
            "type": "object",
            "properties": {
                "authenticated": {
                    "type": "boolean"
                },
                "bom-ref": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.DataClassification"
                    }
                },
                "description": {
                    "type": "string"
                },
                "endpoints": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "externalReferences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.ExternalReference"
                    }
                },
                "group": {
                    "type": "string"
                },
                "licenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.LicenseChoice"
                    }
                },
                "name": {
                    "type": "string"
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Property"
                    }
                },
                "provider": {
                    "$ref": "#/definitions/cyclonedx.OrganizationalEntity"
                },
                "releaseNotes": {
                    "$ref": "#/definitions/cyclonedx.ReleaseNotes"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Service"
                    }
                },
                "version": {
                    "type": "string"
                },
                "x-trust-boundary": {
                    "type": "boolean"
                }
            }
        },
...
        "cyclonedx.SpecVersion": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7
            ],
            "x-enum-comments": {
                "SpecVersion1_0": "1.0",
                "SpecVersion1_1": "1.1",
                "SpecVersion1_2": "1.2",
                "SpecVersion1_3": "1.3",
                "SpecVersion1_4": "1.4",
                "SpecVersion1_5": "1.5",
                "SpecVersion1_6": "1.6"
            },
            "x-enum-varnames": [
                "SpecVersion1_0",
                "SpecVersion1_1",
                "SpecVersion1_2",
                "SpecVersion1_3",
                "SpecVersion1_4",
                "SpecVersion1_5",
                "SpecVersion1_6"
            ]
        },
...
        "cyclonedx.ToolsChoice": {
            "type": "object"
        },
...
        "cyclonedx.Vulnerability": {
            "type": "object",
            "properties": {
                "advisories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Advisory"
                    }
                },
                "affects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Affects"
                    }
                },
                "analysis": {
                    "$ref": "#/definitions/cyclonedx.VulnerabilityAnalysis"
                },
                "bom-ref": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "credits": {
                    "$ref": "#/definitions/cyclonedx.Credits"
                },
                "cwes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "proofOfConcept": {
                    "$ref": "#/definitions/cyclonedx.ProofOfConcept"
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.Property"
                    }
                },
                "published": {
                    "type": "string"
                },
                "ratings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.VulnerabilityRating"
                    }
                },
                "recommendation": {
                    "type": "string"
                },
                "references": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cyclonedx.VulnerabilityReference"
                    }
                },
                "rejected": {
                    "type": "string"
                },
                "source": {
                    "$ref": "#/definitions/cyclonedx.Source"
                },
                "tools": {
                    "$ref": "#/definitions/cyclonedx.ToolsChoice"
                },
                "updated": {
                    "type": "string"
                },
                "workaround": {
                    "type": "string"
                }
            }
        },
...
        "server.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "server.imageGetRequest": {
            "type": "object",
            "required": [
                "image"
            ],
            "properties": {
                "bom": {
                    "$ref": "#/definitions/cyclonedx.BOM"
                },
                "image": {
                    "type": "string",
                    "example": "nginx"
                },
                "plugins": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "certificates"
                    ]
                }
            }
        }
    }
}